Only force symlink creation when overwriting a symlink (#59)
[#173481689]

Authored-by: Tingfang Bao <baotingfang@gmail.com>

+++ ci/concourse/scripts/greenplum-db-6.spec
@@ -89,9 +89,13 @@ exit 0
 %config(noreplace) %{prefix}/greenplum-db-%{gpdb_version}/greenplum_path.sh
 
 %post
-ln -fsT "${RPM_INSTALL_PREFIX}/greenplum-db-%{gpdb_version}" "${RPM_INSTALL_PREFIX}/greenplum-db" || :
-# Set GPHOME to the installation prefix
-sed -i -e "1 s~^\(GPHOME=\).*~\1${RPM_INSTALL_PREFIX}/greenplum-db-%{gpdb_version}~" "${RPM_INSTALL_PREFIX}/greenplum-db-%{gpdb_version}/greenplum_path.sh"
+if [ ! -e "${RPM_INSTALL_PREFIX}/greenplum-db" ] || [ -L "${RPM_INSTALL_PREFIX}/greenplum-db" ];then
+  ln -fsT "${RPM_INSTALL_PREFIX}/greenplum-db-%{gpdb_version}" "${RPM_INSTALL_PREFIX}/greenplum-db" || :
+  # Set GPHOME to the installation prefix
+  sed -i -e "1 s~^\(GPHOME=\).*~\1${RPM_INSTALL_PREFIX}/greenplum-db-%{gpdb_version}~" "${RPM_INSTALL_PREFIX}/greenplum-db-%{gpdb_version}/greenplum_path.sh"
+else
+  echo "the expected symlink was not created because a file exists at that location"
+fi
 
 %postun
 if [ "$(readlink -f "${RPM_INSTALL_PREFIX}/greenplum-db")" == "${RPM_INSTALL_PREFIX}/greenplum-db-%{gpdb_version}" ]; then
